Give context evaluators their own files
